pyproject.toml
[project]
name = "license-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "groq>=1.0.0",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
